change the layout
